+++ examples/minimal/boot.js
@@ -31,7 +31,7 @@ import 'meteor/unchained:core-pricing/plugins/delivery-swiss-tax';
 import 'meteor/unchained:core-filters/plugins/strict-equal';
 import 'meteor/unchained:core-filters/plugins/local-search';
 import 'meteor/unchained:core-quotations/plugins/manual';
-import 'meteor/unchained:core-subscriptions/plugins/licensed';
+import 'meteor/unchained:core-enrollments/plugins/licensed';
 import 'meteor/unchained:core-worker/plugins/external';
 import 'meteor/unchained:core-worker/plugins/http-request';
 import 'meteor/unchained:core-worker/plugins/heartbeat';