+++ packages/api/resolvers/queries/subscriptions.js
@@ -1,7 +1,0 @@
-import { log } from 'meteor/unchained:core-logger';
-import { Subscriptions } from 'meteor/unchained:core-subscriptions';
-
-export default function subscription(root, { limit, offset }, { userId }) {
-  log(`query subscriptions: ${limit} ${offset}`, { userId });
-  return Subscriptions.findSubscriptions({ limit, offset });
-}