+++ packages/core-orders/db/orders/schema.js
@@ -36,7 +36,7 @@ Orders.attachSchema(
       countryCode: String,
       paymentId: String,
       deliveryId: String,
-      originSubscriptionId: String,
+      originEnrollmentId: String,
       ...contextFields,
       ...timestampFields,
       ...calculationFields,