+++ packages/core-subscriptions/db/subscriptions/collections.js
@@ -1,5 +1,0 @@
-import { Mongo } from 'meteor/mongo';
-
-export const Subscriptions = new Mongo.Collection('subscriptions');
-
-export default Subscriptions;